XEP-0077 In-Band Registration added.
