+++ aioxmpp/ibr/__init__.py
@@ -21,5 +21,5 @@
 ########################################################################
 
 from .service import RegistrationService, get_registration_fields, register
-from .service import get_query_xso, get_used_fields
-from .xso import Query
+from .service import get_used_fields
+from .xso import Query, get_query_xso
